+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix the GitHub link on the Add Funds screen
+- `hsd` updated to latest pre-mainnet version
+- Record management updated to reflect latest `hsd` changes
+    - All record types except `NS`, `DS`, `GLUE4/6`, `SYNTH4/6`, and `TXT` have been removed
 
 ## [0.1.1] - 2020-01-22